#38
- `http-server/utils`: специфичные утилиты перенесены в preset'ы (major)
